:bug: Fix commonMain implementation to support iOS compilation
