+++ composeApp/src/commonMain/kotlin/com/crosspaste/notification/NotificationManager.kt
@@ -1,7 +1,7 @@
 package com.crosspaste.notification
 
-import com.crosspaste.utils.ControlUtils.equalDebounce
 import com.crosspaste.utils.GlobalCoroutineScope.ioCoroutineDispatcher
+import com.crosspaste.utils.equalDebounce
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.receiveAsFlow
