+++ composeApp/src/desktopMain/kotlin/com/crosspaste/utils/DispatcherUtils.desktop.kt
@@ -1,0 +1,10 @@
+package com.crosspaste.utils
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+
+actual val ioDispatcher: CoroutineDispatcher = Dispatchers.IO
+
+actual val mainDispatcher: CoroutineDispatcher = Dispatchers.Main
+
+actual val cpuDispatcher: CoroutineDispatcher = Dispatchers.Default
